original_application/requirements_windows.txt
sqlalchemy==2.0.27
psycopg2-binary==2.9.9

# Qt Dependencies - 固定版本以确保兼容性
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
pyqt5-plugins==5.15.9.2.3
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2

numpy==1.26.4
pandas==2.2.3
scikit-learn==1.4.1.post1
tensorflow==2.17.0
matplotlib==3.10.0
seaborn==0.13.2
python-dotenv==1.0.1
requests==2.31.0
pytest==8.0.2
pillow==11.0.0
opencv-python-headless==4.9.0.80
tqdm==4.67.1
loguru==0.7.2
flask==3.0.2
flask-cors==4.0.0
pyyaml==6.0.1
cryptography==42.0.5
mne==1.6.1
PyWavelets==1.5.0
statsmodels==0.14.1
python-docx==1.1.2
PyMuPDF==1.23.26
pdf2image==1.17.0
openpyxl==3.1.5
psutil==6.1.1
typing_extensions==4.12.2
scipy==1.15.1
lazy_loader==0.4
decorator==5.1.1
markdown==3.5.2
pdfkit==1.0.0
